Comments and Fitting
I have added extensive comments for each function in hopes to ameliorate the documentation. I also decided to change the optimization scheme for our fits while fitting for the flux from the basic levmar to the neldermead simplex method since we can obtain higher fidelity results that way. The values are much better and seem to be more than worth the longer fit time!